Make table tests have common abstract parent
